+++ lib/puppet/type/keycloak_protocol_mapper.rb
@@ -177,6 +177,18 @@ Manage Keycloak client scope protocol mappers
     end
   end
 
+  newproperty(:introspection_token_claim, boolean: true) do
+    desc 'introspection.token.claim. Default to `true` for `protocol` `openid-connect`.'
+    newvalues(:true, :false)
+    defaultto do
+      if @resource['protocol'] == 'openid-connect'
+        :true
+      else
+        nil
+      end
+    end
+  end
+
   newproperty(:attribute_nameformat) do
     desc 'attribute.nameformat'
     validate do |v|
@@ -235,15 +247,15 @@ Manage Keycloak client scope protocol mappers
           [:name],
           [:resource_name],
           [:client_scope],
-          [:realm]
-        ]
+          [:realm],
+        ],
       ],
       [
         %r{(.*)},
         [
-          [:name]
-        ]
-      ]
+          [:name],
+        ],
+      ],
     ]
   end
 
@@ -253,7 +265,7 @@ Manage Keycloak client scope protocol mappers
       'oidc-full-name-mapper',
       'oidc-group-membership-mapper',
       'oidc-audience-mapper',
-      'oidc-usermodel-attribute-mapper'
+      'oidc-usermodel-attribute-mapper',
     ]
     if self[:protocol] == 'openid-connect' && !openid_connect_types.include?(self[:type]) && self[:type] !~ %r{script-.+}
       raise Puppet::Error, "type #{self[:type]} is not valid for protocol openid-connect"