+++ lib/puppet/type/keycloak_sssd_user_provider.rb
@@ -79,15 +79,15 @@ Manage Keycloak SSSD user providers
         [
           [:name],
           [:resource_name],
-          [:realm]
-        ]
+          [:realm],
+        ],
       ],
       [
         %r{(.*)},
         [
-          [:name]
-        ]
-      ]
+          [:name],
+        ],
+      ],
     ]
   end